jet brains nerd font local insall

+++ helpers/jet-brains-mono.fish
@@ -1,15 +1,17 @@
 #!/usr/bin/fish
 
-mkdir jet-brains-mono
-pushd jet-brains-mono
+if test ! -d ~/.local/share/fonts
+  mkdir jet-brains-mono
+  pushd jet-brains-mono
 
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/JetBrainsMono.zip
-unzip JetBrainsMono.zip
+  wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.3.3/JetBrainsMono.zip
+  unzip JetBrainsMono.zip
 
-mkdir -p ~/.local/share/fonts
-mv *.ttf ~/.local/share/fonts
+  mkdir -p ~/.local/share/fonts
+  mv *.ttf ~/.local/share/fonts
 
-fc-cache -f -v
+  fc-cache -f -v
 
-popd
-rm -rf jet-brains-mono
+  popd
+  rm -rf jet-brains-mono
+end
